+++ ErsatzTV.Application/FFmpegProfiles/Mapper.cs
@@ -17,6 +17,7 @@ internal static class Mapper
             Resolutions.Mapper.ProjectToViewModel(profile.Resolution),
             profile.ScalingBehavior,
             profile.VideoFormat,
+            profile.VideoProfile,
             profile.BitDepth,
             profile.VideoBitrate,
             profile.VideoBufferSize,