+++ ErsatzTV.FFmpeg/Encoder/Qsv/EncoderH264Qsv.cs
@@ -2,13 +2,29 @@
 
 namespace ErsatzTV.FFmpeg.Encoder.Qsv;
 
-public class EncoderH264Qsv : EncoderBase
+public class EncoderH264Qsv(Option<string> maybeVideoProfile) : EncoderBase
 {
     public override string Name => "h264_qsv";
     public override StreamKind Kind => StreamKind.Video;
 
-    public override string[] OutputOptions =>
-        new[] { "-c:v", "h264_qsv", "-low_power", "0", "-look_ahead", "0" };
+    public override string[] OutputOptions
+    {
+        get
+        {
+            foreach (string videoProfile in maybeVideoProfile)
+            {
+                return
+                [
+                    "-c:v", Name,
+                    "-low_power", "0",
+                    "-look_ahead", "0",
+                    "-profile:v", videoProfile.ToLowerInvariant(),
+                ];
+            }
+
+            return ["-c:v", Name, "-low_power", "0", "-look_ahead", "0"];
+        }
+    }
 
     public override FrameState NextState(FrameState currentState) => currentState with
     {