+++ ErsatzTV.FFmpeg/Format/VideoProfile.cs
@@ -3,4 +3,5 @@ namespace ErsatzTV.FFmpeg.Format;
 public static class VideoProfile
 {
     public const string Main = "main";
+    public const string High = "high";
 }