+++ ErsatzTV.Infrastructure.MySql/Migrations/20240422144541_Add_FFmpegProfile_VideoProfile.cs
@@ -1,0 +1,31 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace ErsatzTV.Infrastructure.MySql.Migrations
+{
+    /// <inheritdoc />
+    public partial class Add_FFmpegProfile_VideoProfile : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "VideoProfile",
+                table: "FFmpegProfile",
+                type: "longtext",
+                nullable: true)
+                .Annotation("MySql:CharSet", "utf8mb4");
+            
+            migrationBuilder.Sql("UPDATE FFmpegProfile SET VideoProfile = 'high'");
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "VideoProfile",
+                table: "FFmpegProfile");
+        }
+    }
+}