+++ ErsatzTV/ViewModels/FFmpegProfileEditViewModel.cs
@@ -33,6 +33,7 @@ public class FFmpegProfileEditViewModel
         VideoBitrate = viewModel.VideoBitrate;
         VideoBufferSize = viewModel.VideoBufferSize;
         VideoFormat = viewModel.VideoFormat;
+        VideoProfile = viewModel.VideoProfile;
         BitDepth = viewModel.BitDepth;
     }
 
@@ -56,6 +57,7 @@ public class FFmpegProfileEditViewModel
     public int VideoBitrate { get; set; }
     public int VideoBufferSize { get; set; }
     public FFmpegProfileVideoFormat VideoFormat { get; set; }
+    public string VideoProfile { get; set; }
     public FFmpegProfileBitDepth BitDepth { get; set; }
 
     public CreateFFmpegProfile ToCreate() =>
@@ -69,6 +71,7 @@ public class FFmpegProfileEditViewModel
             Resolution.Id,
             ScalingBehavior,
             VideoFormat,
+            VideoProfile,
             BitDepth,
             VideoBitrate,
             VideoBufferSize,
@@ -94,6 +97,7 @@ public class FFmpegProfileEditViewModel
             Resolution.Id,
             ScalingBehavior,
             VideoFormat,
+            VideoProfile,
             BitDepth,
             VideoBitrate,
             VideoBufferSize,